He modificado las que estaban mal y añadido libft.h

+++ ft_strlen.c
@@ -3,20 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: natvazqu <natvazqu@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: natvazqu <natvazqu@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 17:35:36 by natvazqu          #+#    #+#             */
-/*   Updated: 2023/12/01 01:20:26 by natvazqu         ###   ########.fr       */
+/*   Updated: 2023/09/12 17:47:36 by natvazqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (s[count] != '\0')
 		count++;
 	return (count);
 }
-// cambio en la condición del while un ´\0´ por un ´0´
